refactor(chat): extract buildA2uiActionMessage to dedicated file
