+++ libs/chat/src/lib/a2ui/surface.component.ts
@@ -6,32 +6,7 @@ import type { A2uiSurface, A2uiActionMessage } from '@cacheplane/a2ui';
 import { RenderSpecComponent, toRenderRegistry } from '@cacheplane/render';
 import type { ViewRegistry, RenderEvent } from '@cacheplane/render';
 import { surfaceToSpec } from './surface-to-spec';
-
-/** Builds a v0.9 A2uiActionMessage from handler params and the current surface. */
-export function buildA2uiActionMessage(
-  params: Record<string, unknown>,
-  surface: A2uiSurface,
-): A2uiActionMessage {
-  const message: A2uiActionMessage = {
-    version: 'v0.9',
-    action: {
-      name: params['name'] as string,
-      surfaceId: surface.surfaceId,
-      sourceComponentId: params['sourceComponentId'] as string,
-      timestamp: new Date().toISOString(),
-      context: (params['context'] as Record<string, unknown>) ?? {},
-    },
-  };
-  if (surface.sendDataModel) {
-    message.metadata = {
-      a2uiClientDataModel: {
-        version: 'v0.9',
-        surfaces: { [surface.surfaceId]: surface.dataModel },
-      },
-    };
-  }
-  return message;
-}
+import { buildA2uiActionMessage } from './build-action-message';
 
 @Component({
   selector: 'a2ui-surface',
