+++ libs/chat/src/public-api.ts
@@ -68,7 +68,7 @@ export type { A2uiSurfaceStore } from './lib/a2ui/surface-store';
 export { A2uiSurfaceComponent } from './lib/a2ui/surface.component';
 export { a2uiBasicCatalog } from './lib/a2ui/catalog/index';
 export { A2uiValidationErrorsComponent } from './lib/a2ui/catalog/validation-errors.component';
-export { buildA2uiActionMessage } from './lib/a2ui/surface.component';
+export { buildA2uiActionMessage } from './lib/a2ui/build-action-message';
 export { surfaceToSpec } from './lib/a2ui/surface-to-spec';
 export type { A2uiActionMessage, A2uiClientDataModel } from '@cacheplane/a2ui';
